requirements.txt
requests==2.24.0
numpy==1.19.5
pandas==1.2.0
tensorflow==2.4.0
scikit-learn==0.24.0
pillow==8.0.1
opencv==4.5.0
flask==1.1.2
dash==1.18.1
dash-core-components==1.14.1
dash-html-components==1.1.1
dash-bootstrap-components==0.11.1
plotly-express==0.4.1